DataTypes

Documentation

data Expr

Data type for expression in a SQL query representing all types of valid expressions

Constructors

IntValue Integer

FloatValue Double

StringValue String

CharValue Char

BoolValue Bool

Null

SQLFunction Aggregate

NestedSelect SQLQuery

Column ColumnRecord

Exists Expr

NotExists Expr

GtAny Expr Expr

GteAny Expr Expr

LteAny Expr Expr

LtAny Expr Expr

EqAny Expr Expr

GtAll Expr Expr

GteAll Expr Expr

LteAll Expr Expr

LtAll Expr Expr

EqAll Expr Expr

NotEq Expr Expr

NotEqAny Expr Expr

NotEqAll Expr Expr

Pow Expr Expr

Mul Expr Expr

Div Expr Expr

Mod Expr Expr

Plus Expr Expr

Minus Expr Expr

Lt Expr Expr

Gt Expr Expr

Lte Expr Expr

Gte Expr Expr

Eq Expr Expr

Is Expr Expr

And Expr Expr

Or Expr Expr

Pos Expr

Neg Expr

Not Expr

data Aggregate

Data type for SQL aggregate function

Constructors

Aggregate

Fields

aggregatefunction :: AggregateFunctions

aggregatecolumn :: String

data ColumnRecord

Data type for a Column record

Constructors

ColumnRecord

Fields

colId :: Maybe String

colName :: String

data ColumnList

Data type for a list of columns. An SQL query may contain column names as well as aggregate functions in select statement so it contains a record having a list of ColumnRecord and Aggregate

Constructors

ColumnList

Fields

columnlist :: [ColumnRecord]

aggregate :: Maybe Aggregate

data Group

Data type for Groupby clause. A Groupby has some column names (ColumnRecord) and an optional having Expr

Constructors

GroupBy

Fields

groupclause :: ColumnRecord

having :: Maybe Expr

data Order

Data type for Orderby clause. An Orderby is followed by the list of column names(ColumnRecord) and the order(asc ordesc)

Constructors

OrderBy

Fields

orderByColumn :: [ColumnRecord]

order :: String

data AggregateFunctions

Data type for aggregate functions ( Min, Max, Count, Avg, Sum)

Constructors

Max

Min

Count

Avg

Sum

data SQLQuery

Data type for an SQL query ( Select, Insert, Delete, Drop, Create, Update)

Constructors

Select

Fields

select :: Either Star ColumnList

from :: [TableRecord]

wherecondition :: Maybe Expr

groupby :: Maybe Group

orderby :: Maybe Order

Insert

Fields

insTable :: TableRecord

insColumns :: Maybe [ColumnRecord]

values :: [Expr]

Delete

Fields

delTable :: TableRecord

delWherecondition :: Expr

Drop

Fields

dropTable :: [TableRecord]

Create

Fields

createTable :: TableRecord

createColumns :: [ColumnPairs]

Update

Fields

updTable :: TableRecord

set :: [(ColumnRecord, Expr)]

updWherecondition :: Maybe Expr

data Star

Data type for selecting everything from table as * in query select * from tablename;

Constructors

Everything

data ColumnPairs

Pairs of column name and the columntype (ColumnTyp) as given in create table command create table tabname (colm1 type1,col2 type2);

Constructors

ColumnPairs String ColumnTyp

data ColumnTyp

Valid types for a column in a table (Int,Double,String,Char,Bool)

Constructors

IntTyp

DoubleTyp

StringTyp

CharTyp

BoolTyp

data TableRecord

Data type for a Table record

Constructors

TableRecord

Fields

tabName :: String

tabId :: Maybe String

sqlDef :: LanguageDef st

Defining all types of tokens that can appear in an input query

sql_identifier :: ParsecT String t Identity String

This is a record containing various parsers for the tokens defined in sqlDef

opTable :: [[Operator String u Identity Expr]]

This is Operator Precedance Table defined for all the operators used in SQL (as defined in sqlDef)