Drawbox

Description

Draws the required graphics on the screen

Documentation

draw :: VertexComponent a => String -> [(a, a, a)] -> GLfloat -> GLfloat -> IO ()

This function draws all the quads and the corresponding strings

points :: [(GLfloat, GLfloat, GLfloat)]

These are the initial points for the first quad

drawError :: String -> IO ()

This funtion prints the error in the input query on screen

drawHeading :: IO ()

This function writes the heading of the output window

drawlines :: VertexComponent a => [(a, a, a)] -> IO ()

This function takes points as input and draws the corresponding line.

drawNothings :: IO ()

drawNothing :: IO ()

drawLineLoop :: VertexComponent a => [(a, a, a)] -> IO ()

This function draws the arrows between quads

drawArrow :: VertexComponent a => [(a, a, a)] -> IO ()

This function draws the arrows from nested query boxes to the outer query