ExecuteQuery

Description

Documentation

construct :: [Char] -> [Char]

To remove parenthesis in aggregate function.

change :: [Char] -> Either ParseError [SQLQuery]

Checks if query contains aggregate function and calls sql_query.

change1 :: [[Char]] -> [[Char]]

Separating aggregate function.

fileop :: String -> IO ()

Functions that takes query; calls sql_query or change and writes the execution order by calling querytype in the file.

querytype :: [Char] -> IO ()

Gets type of query.

writeError :: ParseError -> IO ()

Writes parse Error to file.

selectquery :: String -> IO ()

Checks if SELECT query has any,all,exists and calls anyall ,existquery, groupquery and selectsql to write execution order to file.

anyallnest :: [Char] -> [String]

Helping function for anyall

anyall :: String -> IO ()

Execution order for nested SELECT querries with any, all calls anyallnest and arrange.

existquery :: String -> IO ()

Execution order for nested SELECT querries with EXISTS or NOT EXISTS.

arrange :: [a] -> [a]

Helping function for getting Execution order.

selectsql :: String -> IO ()

Execution order for SELECT query calls fun

fun :: [String] -> [String] -> [String]

Helper function for selectsql

removecomma :: [Char] -> [Char]

Functions to check the constraint : columns inside ORDER BY is part of columns inside SELECT.

orderelement :: [[Char]] -> [[Char]] -> Bool

Checks constraint : columns in ORDER BY is part of columns inside SELECT.

columnlist :: [Char] -> [[Char]]

Finds list of columns inside SELECT.

orderlist :: [Char] -> [[Char]]

Finds list of columns inside ORDER BY.

ordersyntax :: [Char] -> IO ()

Checks ORDER BY constraint by calling orderelement and writes execution order to file.

breakgroup :: String -> [String]

Breaks GROUP BY querries.

groupquery :: String -> IO ()

Writes execution order for GROUP BY query.

removebracket :: [[Char]] -> [[Char]]

Function to remove parenthesis from aggregate function.

insertsyntax :: [Char] -> IO ()

Checks if values in INSERT query is same as number of columns; using removebracket.

callSqlParser :: String -> IO ()

Checks for SELECT or INSERT query and accordingly calls fileop

sqlQuery :: String -> IO ()

Main function which instanciate function like callSqlParser and display