ReadnSend

Description

Takes a list as an arg and calls draw for each token

Documentation

breaklist :: [String] -> ([String], [String])

This function breaks the list of tokens from input file when it finds a nested query

fst' :: (t, t1) -> t

This function takes a tuple of lists and returns its first list

snd' :: (t, t1) -> t1

This function takes a tuple of lists and returns its second list

checkForError :: [[Char]] -> Bool

This function checks if the input tokens are valid or not

readNsend :: [[Char]] -> IO ()

This function calls drawup , drawlow , drawHeadingor drawError depending on the tokens in the input list

drawup :: Eq a => [String] -> [a] -> Int -> IO ()

This function draws the graphics for nested query

outerBoxPts :: Int -> [(GLfloat, GLfloat, GLfloat)]

arrowPts :: Int -> [(GLfloat, GLfloat, GLfloat)] -> [(GLfloat, GLfloat, GLfloat)]

This function is used for calculating points for drawing the arrow graphics

drawlow :: [String] -> Int -> IO ()

This function draws the graphics for outer query

upd :: (Fractional a, Integral a1, VertexComponent a) => String -> [(a, a, a)] -> a1 -> GLfloat -> GLfloat -> IO ()

This function calculates new points for drawing box graphics and calls draw function

x :: GLfloat

yu :: GLfloat

yd :: GLfloat

linepu :: [(GLfloat, GLfloat, GLfloat)]

linepd :: [(GLfloat, GLfloat, GLfloat)]

ptsArrowDwn :: [(GLfloat, GLfloat, GLfloat)]

pointsu :: [(GLfloat, GLfloat, GLfloat)]

pointsl :: [(GLfloat, GLfloat, GLfloat)]

add :: (Fractional t, Integral a) => a -> (t, t1, t2) -> (t, t1, t2)

This function updates the points for the next box whenever a new token is encountered

getlinepts :: (Eq a, Fractional t, Num a) => a -> (GLfloat, t, t1) -> [(GLfloat, t, t1)]

This function calculates the points for the arrows between two boxes

head' :: [a] -> a

This functions returns the element at the first position of the list