SQLParser

Documentation

sql_table :: ParsecT String t Identity TableRecord

Parser for parsing names of tables

sql_columns :: Parser ColumnRecord

Parser for parsing ColumnRecord and checking valid format for a column name.

sql_ColumnRecord :: Parser ColumnList

Parser for parsing ColumnList and checking the column names format and the aggregate functions as well

sql_aggregate :: ParsecT String () Identity Aggregate

Parser for Aggregate functions.

sql_agfunctions :: Parser AggregateFunctions

Parser for Aggregate functions.

sql_Expr :: Parser Expr

Parser for parsing the expressions in an SQL query statement

sql_drop :: Parser SQLQuery

Parser for SQL DROP query

sql_delete :: Parser SQLQuery

Parser for SQL DELETE query

sql_update :: Parser SQLQuery

Parser for SQL UPDATE query

sql_set :: Parser (ColumnRecord, Expr)

Parser for SET inside UPDATE query

sql_create :: Parser SQLQuery

Parser for SQL CREATE query

sql_createcolumn :: Parser ColumnPairs

Parser for creating columns with data types inside CREATE query

sql_columntype :: Parser ColumnTyp

Parser for SQL column types

sql_insert :: Parser SQLQuery

Parser for SQL INSERT query

sql_select :: Parser SQLQuery

Parser for SQL SELECT query

sql_group :: ParsecT String () Identity Group

Parser for SQL GROUP BY uses sql_havingExpr

sql_havingExpr :: Parser Expr

Parser for SQL HAVING clause

sql_sqlStatement :: Parser SQLQuery

Parser that calls parser for different SQL querries.

sql_query :: String -> Either ParseError [SQLQuery]

Parses SQL statements